requirements.txt
PyYAML
requests
IPy
delegator.py
paramiko